First working copy
Somewhat major changing mostly including a change from a bash script into a golang TUI.
This means that golang will be used to produce a TUI application instead. This is just a first copy with a draft version of
a gui in which there currently is no real functioning product yet.